refactor: Make Discovery a DataClass.
And move it out of the fsm.py file.
